+++ libraries/AS5600/AS5600.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: AS56000.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.3.5
+// VERSION: 0.3.6
 // PURPOSE: Arduino library for AS5600 magnetic rotation meter
 //    DATE: 2022-05-28
 //     URL: https://github.com/RobTillaart/AS5600
@@ -451,18 +451,19 @@ float AS5600::getAngularSpeed(uint8_t mode)
 //
 //  POSITION cumulative
 //
-
 int32_t AS5600::getCumulativePosition()
 {
   int16_t value = readReg2(AS5600_RAW_ANGLE) & 0x0FFF;
 
   //  whole rotation CW?
-  if ((_lastPosition > 2048) && ( value < (_lastPosition - 2048)))  //  less than half a circle
+  //  less than half a circle
+  if ((_lastPosition > 2048) && ( value < (_lastPosition - 2048)))
   {
     _position = _position + 4096 - _lastPosition + value;
   }
   //  whole rotation CCW?
-  else if ((value > 2048) && ( _lastPosition < (value - 2048)))  //  less than half a circle
+  //  less than half a circle
+  else if ((value > 2048) && ( _lastPosition < (value - 2048)))
   {
     _position = _position - 4096 - _lastPosition + value;
   }
@@ -475,17 +476,26 @@ int32_t AS5600::getCumulativePosition()
 
 int32_t AS5600::getRevolutions()
 {
-  int32_t p =  _position >> 12;
+  int32_t p = _position >> 12;  //  divide by 4096
   return p;
   // if (p < 0) p++;
   // return p;
 }
 
 
-int32_t AS5600::resetPosition()
+int32_t AS5600::resetPosition(int32_t position)
 {
   int32_t old = _position;
-  _position  = 0;
+  _position = position;
+  return old;
+}
+
+
+int32_t AS5600::resetCumulativePosition(int32_t position)
+{
+  _lastPosition = readReg2(AS5600_RAW_ANGLE) & 0x0FFF;
+  int32_t old = _position;
+  _position = position;
   return old;
 }